UI: changed helper to float values

+++ src/modules/ui/imgui/IMGUI.cpp
@@ -158,8 +158,8 @@ bool DisabledButton(const char *text, bool disabled) {
 	return pressed;
 }
 
-int Size(int size) {
-	return (int)((float)size * ImGui::GetWindowDpiScale());
+float Size(float size) {
+	return size * ImGui::GetWindowDpiScale();
 }
 
 }
